// frontend/generated/models/ToyDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 8lind8box Platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ImageDto } from './ImageDto';
import {
    ImageDtoFromJSON,
    ImageDtoFromJSONTyped,
    ImageDtoToJSON,
    ImageDtoToJSONTyped,
} from './ImageDto';

/**
 * 
 * @export
 * @interface ToyDto
 */
export interface ToyDto {
    /**
     * The unique identifier for a toy
     * @type {number}
     * @memberof ToyDto
     */
    toyId?: number;
    /**
     * The name of the toy
     * @type {string}
     * @memberof ToyDto
     */
    name?: string;
    /**
     * The description of the toy
     * @type {string}
     * @memberof ToyDto
     */
    description?: string;
    /**
     * The weight of the toy
     * @type {number}
     * @memberof ToyDto
     */
    weight?: number;
    /**
     * The rarity of the toy
     * @type {string}
     * @memberof ToyDto
     */
    rarity?: ToyDtoRarityEnum;
    /**
     * Visibility status of the toy
     * @type {boolean}
     * @memberof ToyDto
     */
    isVisible?: boolean;
    /**
     * The creation timestamp of the toy
     * @type {Date}
     * @memberof ToyDto
     */
    createdAt?: Date;
    /**
     * The last update timestamp of the toy
     * @type {Date}
     * @memberof ToyDto
     */
    updatedAt?: Date;
    /**
     * 
     * @type {Array<ImageDto>}
     * @memberof ToyDto
     */
    images?: Array<ImageDto>;
}


/**
 * @export
 */
export const ToyDtoRarityEnum = {
    Regular: 'REGULAR',
    Secret: 'SECRET'
} as const;
export type ToyDtoRarityEnum = typeof ToyDtoRarityEnum[keyof typeof ToyDtoRarityEnum];


/**
 * Check if a given object implements the ToyDto interface.
 */
export function instanceOfToyDto(value: object): value is ToyDto {
    return true;
}

export function ToyDtoFromJSON(json: any): ToyDto {
    return ToyDtoFromJSONTyped(json, false);
}

export function ToyDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ToyDto {
    if (json == null) {
        return json;
    }
    return {
        
        'toyId': json['toyId'] == null ? undefined : json['toyId'],
        'name': json['name'] == null ? undefined : json['name'],
        'description': json['description'] == null ? undefined : json['description'],
        'weight': json['weight'] == null ? undefined : json['weight'],
        'rarity': json['rarity'] == null ? undefined : json['rarity'],
        'isVisible': json['isVisible'] == null ? undefined : json['isVisible'],
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
        'updatedAt': json['updatedAt'] == null ? undefined : (new Date(json['updatedAt'])),
        'images': json['images'] == null ? undefined : ((json['images'] as Array<any>).map(ImageDtoFromJSON)),
    };
}

export function ToyDtoToJSON(json: any): ToyDto {
    return ToyDtoToJSONTyped(json, false);
}

export function ToyDtoToJSONTyped(value?: ToyDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'toyId': value['toyId'],
        'name': value['name'],
        'description': value['description'],
        'weight': value['weight'],
        'rarity': value['rarity'],
        'isVisible': value['isVisible'],
        'createdAt': value['createdAt'] == null ? undefined : ((value['createdAt']).toISOString()),
        'updatedAt': value['updatedAt'] == null ? undefined : ((value['updatedAt']).toISOString()),
        'images': value['images'] == null ? undefined : ((value['images'] as Array<any>).map(ImageDtoToJSON)),
    };
}